Move content to components
